Restructure some tests into classes, add repr(Object3d) test
